refactor: code organisation and documentation details
